Changing planning horizon to double

+++ src/repastInterSim/agent/Ped.java
@@ -64,7 +64,7 @@ public class Ped extends MobileAgent {
 
     private HashMap<Integer, Double> gridSummandPriorityMap = new HashMap<Integer, Double>(); // Used to get grid cell summand value when running flood fill algorithm for routing
     
-    private Integer pHorizon = 20; // Tactical planning horizon of ped agent in degrees
+    private Double pHorizon = 20.0; // Tactical planning horizon of ped agent in degrees
     
     private double vehiclePriorityCostRatio; // The ratio of pedestrian priority cell cost to vehicle priority cell cost. Represents pedestrian's perception of cost of moving in vehicle priority space.
     
@@ -691,12 +691,12 @@ public class Ped extends MobileAgent {
     }
 
 
-	public Integer getpHorizon() {
+	public Double getpHorizon() {
 		return pHorizon;
 	}
 
 
-	public void setpHorizon(Integer pHorizon) {
+	public void setpHorizon(Double pHorizon) {
 		this.pHorizon = pHorizon;
 	}
